The weather banner is displayed. Now need to make it work both landscape and portrait. Need to make sure the image fits the screen.
git-svn-id: https://192.168.2.108/repos/BbApps/trunk/BB_Development/EclipseProjects/MusicalClock@274 6d98d622-1d27-443a-aa54-b1e34a34e7a4

+++ src/com/karvitech/api/weather/UpdateWeatherRunnable.java
@@ -6,7 +6,7 @@ public class UpdateWeatherRunnable implements Runnable{
 	private String mUrl;
 	private WeatherInfoListener mListener;
 	
-	public UpdateWeatherRunnable(float lat, float longi, WeatherInfoListener listener) {
+	public UpdateWeatherRunnable(double lat, double longi, WeatherInfoListener listener) {
 		mUrl = WeatherUtils.YR_NO_FORECAST_URL + "lat=" + lat + ";lon=" + longi;
 		mListener = listener;
 	}
